pyproject.toml
[project]
name = "ordra-ai"
version = "0.1.0"
description = "ORDRA-AI: Agentic Order-to-Cash Intake for SAP (capstone)"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
  # API
  "fastapi>=0.110,<0.116",
  "uvicorn[standard]>=0.27,<0.32",
  "pydantic>=2.6,<2.11",
  "python-dotenv>=1.0,<2.0",

  # UI
  "streamlit>=1.30,<1.40",
  "requests>=2.31,<3.0",

  # Orchestration / graph
  "networkx>=3.2,<3.4",
  "pyyaml>=6.0,<7.0",

  # LLM / templating / schema validation
  "openai>=1.0,<2.0",
  "jinja2>=3.1,<3.2",
  "jsonschema>=4.21,<4.24",

  # Vector memory
  "numpy>=1.26,<3.0",
  "faiss-cpu>=1.7.4,<1.9",

  # Mailbox / docs from bytes
  "pypdf>=4.0,<6.0",

  # Testing
  "pytest>=7.4,<9.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4,<0.7",
  "mypy>=1.8,<1.12"
]
demo = [
  "reportlab>=4.0,<5.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools.packages.find]
include = ["ordra*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
